Actually fixed the lease agreements - had to nuke the db
