--- backend/app/alembic/versions/0a55b92fc079_changed_conversations_sender_name_to_.py
@@ -1,32 +1,0 @@
-"""Changed conversations sender_name to name
-
-Revision ID: 0a55b92fc079
-Revises: 344d7394747f
-Create Date: 2025-03-30 04:04:11.357644
-
-"""
-from typing import Sequence, Union
-import sqlmodel.sql.sqltypes
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '0a55b92fc079'
-down_revision: Union[str, None] = '344d7394747f'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('conversation', sa.Column('name', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=True))
-    op.drop_column('conversation', 'sender_name')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('conversation', sa.Column('sender_name', sa.VARCHAR(length=255), autoincrement=False, nullable=True))
-    op.drop_column('conversation', 'name')
-    # ### end Alembic commands ###
