--- backend/app/alembic/versions/239fac48ae6a_add_column_phone_number_to_user.py
@@ -1,32 +1,0 @@
-"""Add column phone_number to user
-
-Revision ID: 239fac48ae6a
-Revises: 1a31ce608336
-Create Date: 2025-02-27 13:14:38.833167
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '239fac48ae6a'
-down_revision: Union[str, None] = '1a31ce608336'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('user', sa.Column('phone', sa.String(length=10), nullable=True))
-    op.create_index(op.f('ix_user_phone'), 'user', ['phone'], unique=True)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_user_phone'), table_name='user')
-    op.drop_column('user', 'phone')
-    # ### end Alembic commands ###
