--- backend/app/alembic/versions/29cafea606b9_add_column_bio_and_profile_type_to_user_.py
@@ -1,32 +1,0 @@
-"""Add column bio and profile_type to User model
-
-Revision ID: 29cafea606b9
-Revises: c1cfcf9de813
-Create Date: 2025-03-27 20:25:29.758546
-
-"""
-from typing import Sequence, Union
-import sqlmodel.sql.sqltypes
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '29cafea606b9'
-down_revision: Union[str, None] = 'c1cfcf9de813'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('user', sa.Column('bio', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=True))
-    op.add_column('user', sa.Column('profile_type', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('user', 'profile_type')
-    op.drop_column('user', 'bio')
-    # ### end Alembic commands ###
