--- backend/app/alembic/versions/2b86cf182ea1_added_images_models_for_listings.py
@@ -1,41 +1,0 @@
-"""Added images models for listings
-
-Revision ID: 2b86cf182ea1
-Revises: 29cafea606b9
-Create Date: 2025-04-02 23:28:18.813408
-
-"""
-from typing import Sequence, Union
-import sqlmodel.sql.sqltypes
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '2b86cf182ea1'
-down_revision: Union[str, None] = '29cafea606b9'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('image',
-    sa.Column('filename', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=False),
-    sa.Column('file_path', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=False),
-    sa.Column('file_type', sa.Enum('JPEG', 'JPG', 'PNG', 'WEBP', 'GIF', name='filetype'), nullable=False),
-    sa.Column('file_size', sa.Integer(), nullable=False),
-    sa.Column('is_primary', sa.Boolean(), nullable=False),
-    sa.Column('display_order', sa.Integer(), nullable=False),
-    sa.Column('id', sa.Uuid(), nullable=False),
-    sa.Column('listing_id', sa.Uuid(), nullable=False),
-    sa.ForeignKeyConstraint(['listing_id'], ['listing.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('image')
-    # ### end Alembic commands ###
