--- backend/app/alembic/versions/34565e904379_changed_autologout_time_to_a_float_so_.py
@@ -1,36 +1,0 @@
-"""changed autologout time to a float so we can have it in seconds for project demo
-
-Revision ID: 34565e904379
-Revises: df28aa0800de
-Create Date: 2025-03-05 01:51:35.886975
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '34565e904379'
-down_revision: Union[str, None] = 'df28aa0800de'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('user', 'auto_logout',
-               existing_type=sa.INTEGER(),
-               type_=sa.Float(),
-               existing_nullable=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('user', 'auto_logout',
-               existing_type=sa.Float(),
-               type_=sa.INTEGER(),
-               existing_nullable=False)
-    # ### end Alembic commands ###
