--- backend/app/alembic/versions/436911304233_image_file_type.py
@@ -1,30 +1,0 @@
-"""Image file type
-
-Revision ID: 436911304233
-Revises: 8b348c51a4f5
-Create Date: 2025-04-03 21:58:41.118862
-
-"""
-from typing import Sequence, Union
-import sqlmodel.sql.sqltypes
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '436911304233'
-down_revision: Union[str, None] = '8b348c51a4f5'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
