--- backend/app/alembic/versions/4b3272ada45a_add_column_pin_to_users.py
@@ -1,30 +1,0 @@
-"""Add column pin to users
-
-Revision ID: 4b3272ada45a
-Revises: 98195a5a2b1c
-Create Date: 2025-03-02 20:25:47.772786
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '4b3272ada45a'
-down_revision: Union[str, None] = '98195a5a2b1c'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
