--- backend/app/alembic/versions/660ab45a966f_add_saved_listings_to_userbase.py
@@ -1,30 +1,0 @@
-"""Add saved_listings to UserBase
-
-Revision ID: 660ab45a966f
-Revises: 9a3475bcc340
-Create Date: 2025-04-01 02:25:49.898240
-
-"""
-from typing import Sequence, Union
-import sqlmodel.sql.sqltypes
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '660ab45a966f'
-down_revision: Union[str, None] = '9a3475bcc340'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
