--- backend/app/alembic/versions/66e7fab59f8a_adding_pfds.py
@@ -1,40 +1,0 @@
-"""Adding pfds
-
-Revision ID: 66e7fab59f8a
-Revises: d4833295feb2
-Create Date: 2025-04-03 20:46:00.004651
-
-"""
-from typing import Sequence, Union
-import sqlmodel.sql.sqltypes
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '66e7fab59f8a'
-down_revision: Union[str, None] = 'd4833295feb2'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('leaseagreement',
-    sa.Column('filename', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=False),
-    sa.Column('file_path', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=False),
-    sa.Column('file_type', sa.Enum('PDF', 'TXT', name='leasefiletype'), nullable=False),
-    sa.Column('file_size', sa.Integer(), nullable=False),
-    sa.Column('description', sqlmodel.sql.sqltypes.AutoString(length=500), nullable=True),
-    sa.Column('id', sa.Uuid(), nullable=False),
-    sa.Column('listing_id', sa.Uuid(), nullable=False),
-    sa.ForeignKeyConstraint(['listing_id'], ['listing.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('leaseagreement')
-    # ### end Alembic commands ###
