--- backend/app/alembic/versions/727bd0825888_add_messaging_table_and_blocking_rules_.py
@@ -1,73 +1,0 @@
-"""Add messaging table and blocking rules to user
-
-Revision ID: 727bd0825888
-Revises: 9d2c7eb6f434
-Create Date: 2025-03-07 00:41:27.135468
-
-"""
-from typing import Sequence, Union
-import sqlmodel.sql.sqltypes
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '727bd0825888'
-down_revision: Union[str, None] = '9d2c7eb6f434'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('conversation',
-    sa.Column('sender_name', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=True),
-    sa.Column('is_group', sa.Boolean(), nullable=False),
-    sa.Column('id', sa.Uuid(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('conversationparticipant',
-    sa.Column('conversation_id', sa.Uuid(), nullable=False),
-    sa.Column('user_id', sa.Uuid(), nullable=False),
-    sa.ForeignKeyConstraint(['conversation_id'], ['conversation.id'], ),
-    sa.ForeignKeyConstraint(['user_id'], ['user.id'], ),
-    sa.PrimaryKeyConstraint('conversation_id', 'user_id')
-    )
-    op.create_table('userblock',
-    sa.Column('blocker_id', sa.Uuid(), nullable=False),
-    sa.Column('blocked_id', sa.Uuid(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.ForeignKeyConstraint(['blocked_id'], ['user.id'], ),
-    sa.ForeignKeyConstraint(['blocker_id'], ['user.id'], ),
-    sa.PrimaryKeyConstraint('blocker_id', 'blocked_id')
-    )
-    op.drop_table('messagerecipient')
-    op.add_column('message', sa.Column('conversation_id', sa.Uuid(), nullable=False))
-    op.add_column('message', sa.Column('updated_at', sa.DateTime(), nullable=True))
-    op.add_column('message', sa.Column('deleted', sa.Boolean(), nullable=False))
-    op.add_column('message', sa.Column('deleted_at', sa.DateTime(), nullable=True))
-    op.create_foreign_key(None, 'message', 'conversation', ['conversation_id'], ['id'])
-    op.drop_column('message', 'is_group_chat')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('message', sa.Column('is_group_chat', sa.BOOLEAN(), autoincrement=False, nullable=False))
-    op.drop_constraint(None, 'message', type_='foreignkey')
-    op.drop_column('message', 'deleted_at')
-    op.drop_column('message', 'deleted')
-    op.drop_column('message', 'updated_at')
-    op.drop_column('message', 'conversation_id')
-    op.create_table('messagerecipient',
-    sa.Column('message_id', sa.UUID(), autoincrement=False, nullable=False),
-    sa.Column('user_id', sa.UUID(), autoincrement=False, nullable=False),
-    sa.ForeignKeyConstraint(['message_id'], ['message.id'], name='messagerecipient_message_id_fkey'),
-    sa.ForeignKeyConstraint(['user_id'], ['user.id'], name='messagerecipient_user_id_fkey'),
-    sa.PrimaryKeyConstraint('message_id', 'user_id', name='messagerecipient_pkey')
-    )
-    op.drop_table('userblock')
-    op.drop_table('conversationparticipant')
-    op.drop_table('conversation')
-    # ### end Alembic commands ###
