--- backend/app/alembic/versions/76d703caa427_renamed_phone_in_user_model.py
@@ -1,36 +1,0 @@
-"""renamed phone in  User model
-
-Revision ID: 76d703caa427
-Revises: 239fac48ae6a
-Create Date: 2025-02-27 13:23:08.828637
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '76d703caa427'
-down_revision: Union[str, None] = '239fac48ae6a'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('user', sa.Column('phone_number', sa.String(length=10), nullable=True))
-    op.drop_index('ix_user_phone', table_name='user')
-    op.create_index(op.f('ix_user_phone_number'), 'user', ['phone_number'], unique=True)
-    op.drop_column('user', 'phone')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('user', sa.Column('phone', sa.VARCHAR(length=10), autoincrement=False, nullable=True))
-    op.drop_index(op.f('ix_user_phone_number'), table_name='user')
-    op.create_index('ix_user_phone', 'user', ['phone'], unique=True)
-    op.drop_column('user', 'phone_number')
-    # ### end Alembic commands ###
