--- backend/app/alembic/versions/86fadaef55d0_removed_unnecessary_columns_from_.py
@@ -1,50 +1,0 @@
-"""Removed unnecessary columns from listings
-
-Revision ID: 86fadaef55d0
-Revises: df4f0bb192ae
-Create Date: 2025-03-15 20:12:43.494830
-
-"""
-from typing import Sequence, Union
-import sqlmodel.sql.sqltypes
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = '86fadaef55d0'
-down_revision: Union[str, None] = 'df4f0bb192ae'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('listing', sa.Column('rent', sa.Float(), nullable=True))
-    op.alter_column('listing', 'num_bedrooms',
-               existing_type=postgresql.JSONB(astext_type=sa.Text()),
-               type_=sqlmodel.sql.sqltypes.AutoString(),
-               existing_nullable=True)
-    op.alter_column('listing', 'num_bathrooms',
-               existing_type=postgresql.JSONB(astext_type=sa.Text()),
-               type_=sqlmodel.sql.sqltypes.AutoString(),
-               existing_nullable=True)
-    op.drop_column('listing', 'max_price')
-    op.drop_column('listing', 'min_price')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('listing', sa.Column('min_price', sa.DOUBLE_PRECISION(precision=53), autoincrement=False, nullable=True))
-    op.add_column('listing', sa.Column('max_price', sa.DOUBLE_PRECISION(precision=53), autoincrement=False, nullable=True))
-    op.alter_column('listing', 'num_bathrooms',
-               existing_type=sqlmodel.sql.sqltypes.AutoString(),
-               type_=postgresql.JSONB(astext_type=sa.Text()),
-               existing_nullable=True)
-    op.alter_column('listing', 'num_bedrooms',
-               existing_type=sqlmodel.sql.sqltypes.AutoString(),
-               type_=postgresql.JSONB(astext_type=sa.Text()),
-               existing_nullable=True)
-    op.drop_column('listing', 'rent')
-    # ### end Alembic commands ###
