--- backend/app/alembic/versions/8b348c51a4f5_lease_agreements_one_to_one.py
@@ -1,30 +1,0 @@
-"""lease agreements one-to-one
-
-Revision ID: 8b348c51a4f5
-Revises: b1be86446a3e
-Create Date: 2025-04-03 21:36:19.704110
-
-"""
-from typing import Sequence, Union
-import sqlmodel.sql.sqltypes
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '8b348c51a4f5'
-down_revision: Union[str, None] = 'b1be86446a3e'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_unique_constraint(None, 'leaseagreement', ['listing_id'])
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'leaseagreement', type_='unique')
-    # ### end Alembic commands ###
