--- backend/app/alembic/versions/950410132acd_add_auto_logout_time_to_user_table.py
@@ -1,30 +1,0 @@
-"""Add auto_logout time to user table
-
-Revision ID: 950410132acd
-Revises: 76d703caa427
-Create Date: 2025-03-03 18:43:13.913730
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '950410132acd'
-down_revision: Union[str, None] = '76d703caa427'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('user', sa.Column('auto_logout', sa.Integer(), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('user', 'auto_logout')
-    # ### end Alembic commands ###
