--- backend/app/alembic/versions/98195a5a2b1c_add_column_hashed_pin_to_user_model.py
@@ -1,31 +1,0 @@
-"""Add column hashed_pin to User model
-
-Revision ID: 98195a5a2b1c
-Revises: 76d703caa427
-Create Date: 2025-03-01 19:55:10.641461
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-import sqlmodel.sql.sqltypes
-
-
-# revision identifiers, used by Alembic.
-revision: str = '98195a5a2b1c'
-down_revision: Union[str, None] = '76d703caa427'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('user', sa.Column('hashed_pin', sqlmodel.sql.sqltypes.AutoString(), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('user', 'hashed_pin')
-    # ### end Alembic commands ###
