--- backend/app/alembic/versions/9d2c7eb6f434_added_tables_for_messaging.py
@@ -1,55 +1,0 @@
-"""Added tables for messaging
-
-Revision ID: 9d2c7eb6f434
-Revises: 1a31ce608336
-Create Date: 2025-03-03 21:31:31.446602
-
-"""
-from typing import Sequence, Union
-import sqlmodel.sql.sqltypes
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '9d2c7eb6f434'
-down_revision: Union[str, None] = '1a31ce608336'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('message',
-    sa.Column('content', sqlmodel.sql.sqltypes.AutoString(length=4000), nullable=False),
-    sa.Column('is_group_chat', sa.Boolean(), nullable=False),
-    sa.Column('id', sa.Uuid(), nullable=False),
-    sa.Column('sender_id', sa.Uuid(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=False),
-    sa.ForeignKeyConstraint(['sender_id'], ['user.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('messagereadstatus',
-    sa.Column('message_id', sa.Uuid(), nullable=False),
-    sa.Column('user_id', sa.Uuid(), nullable=False),
-    sa.Column('read_at', sa.DateTime(), nullable=False),
-    sa.ForeignKeyConstraint(['message_id'], ['message.id'], ),
-    sa.ForeignKeyConstraint(['user_id'], ['user.id'], ),
-    sa.PrimaryKeyConstraint('message_id', 'user_id')
-    )
-    op.create_table('messagerecipient',
-    sa.Column('message_id', sa.Uuid(), nullable=False),
-    sa.Column('user_id', sa.Uuid(), nullable=False),
-    sa.ForeignKeyConstraint(['message_id'], ['message.id'], ),
-    sa.ForeignKeyConstraint(['user_id'], ['user.id'], ),
-    sa.PrimaryKeyConstraint('message_id', 'user_id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('messagerecipient')
-    op.drop_table('messagereadstatus')
-    op.drop_table('message')
-    # ### end Alembic commands ###
