--- backend/app/alembic/versions/c1cfcf9de813_removed_column_and_jsonb_from_listings.py
@@ -1,44 +1,0 @@
-"""removed column and JSONB from listings
-
-Revision ID: c1cfcf9de813
-Revises: 86fadaef55d0
-Create Date: 2025-03-25 19:53:09.526640
-
-"""
-from typing import Sequence, Union
-import sqlmodel.sql.sqltypes
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = 'c1cfcf9de813'
-down_revision: Union[str, None] = '86fadaef55d0'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('listing', 'included_utilities',
-               existing_type=postgresql.JSONB(astext_type=sa.Text()),
-               type_=sa.JSON(),
-               existing_nullable=True)
-    op.alter_column('listing', 'amenities',
-               existing_type=postgresql.JSONB(astext_type=sa.Text()),
-               type_=sa.JSON(),
-               existing_nullable=True)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('listing', 'amenities',
-               existing_type=sa.JSON(),
-               type_=postgresql.JSONB(astext_type=sa.Text()),
-               existing_nullable=True)
-    op.alter_column('listing', 'included_utilities',
-               existing_type=sa.JSON(),
-               type_=postgresql.JSONB(astext_type=sa.Text()),
-               existing_nullable=True)
-    # ### end Alembic commands ###
