+++ backend/app/alembic/versions/cce7fa393846_initial_revision.py
@@ -1,0 +1,150 @@
+"""Initial Revision
+
+Revision ID: cce7fa393846
+Revises: 
+Create Date: 2025-04-03 23:16:39.755705
+
+"""
+from typing import Sequence, Union
+import sqlmodel.sql.sqltypes
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = 'cce7fa393846'
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('conversation',
+    sa.Column('name', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=True),
+    sa.Column('is_group', sa.Boolean(), nullable=False),
+    sa.Column('id', sa.Uuid(), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('user',
+    sa.Column('email', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=False),
+    sa.Column('phone_number', sqlmodel.sql.sqltypes.AutoString(length=10), nullable=True),
+    sa.Column('is_active', sa.Boolean(), nullable=False),
+    sa.Column('is_superuser', sa.Boolean(), nullable=False),
+    sa.Column('full_name', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=True),
+    sa.Column('bio', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=True),
+    sa.Column('profile_type', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=True),
+    sa.Column('auto_logout', sa.Float(), nullable=False),
+    sa.Column('id', sa.Uuid(), nullable=False),
+    sa.Column('hashed_password', sqlmodel.sql.sqltypes.AutoString(), nullable=False),
+    sa.Column('hashed_pin', sqlmodel.sql.sqltypes.AutoString(), nullable=True),
+    sa.Column('is_2fa_enabled', sa.Boolean(), nullable=True),
+    sa.Column('latest_otp', sqlmodel.sql.sqltypes.AutoString(), nullable=True),
+    sa.Column('saved_listings', sa.JSON(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_user_email'), 'user', ['email'], unique=True)
+    op.create_index(op.f('ix_user_phone_number'), 'user', ['phone_number'], unique=True)
+    op.create_table('conversationparticipant',
+    sa.Column('conversation_id', sa.Uuid(), nullable=False),
+    sa.Column('user_id', sa.Uuid(), nullable=False),
+    sa.ForeignKeyConstraint(['conversation_id'], ['conversation.id'], ),
+    sa.ForeignKeyConstraint(['user_id'], ['user.id'], ),
+    sa.PrimaryKeyConstraint('conversation_id', 'user_id')
+    )
+    op.create_table('item',
+    sa.Column('description', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=True),
+    sa.Column('id', sa.Uuid(), nullable=False),
+    sa.Column('title', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=False),
+    sa.Column('owner_id', sa.Uuid(), nullable=False),
+    sa.ForeignKeyConstraint(['owner_id'], ['user.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('listing',
+    sa.Column('num_bedrooms', sqlmodel.sql.sqltypes.AutoString(), nullable=True),
+    sa.Column('num_bathrooms', sqlmodel.sql.sqltypes.AutoString(), nullable=True),
+    sa.Column('address', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=True),
+    sa.Column('realty_company', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=True),
+    sa.Column('rent', sa.Float(), nullable=True),
+    sa.Column('included_utilities', sa.JSON(), nullable=True),
+    sa.Column('security_deposit', sqlmodel.sql.sqltypes.AutoString(), nullable=True),
+    sa.Column('amenities', sa.JSON(), nullable=True),
+    sa.Column('lease_start_date', sqlmodel.sql.sqltypes.AutoString(), nullable=True),
+    sa.Column('lease_end_date', sqlmodel.sql.sqltypes.AutoString(), nullable=True),
+    sa.Column('id', sa.Uuid(), nullable=False),
+    sa.Column('owner_id', sa.Uuid(), nullable=False),
+    sa.ForeignKeyConstraint(['owner_id'], ['user.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('message',
+    sa.Column('content', sqlmodel.sql.sqltypes.AutoString(length=4000), nullable=False),
+    sa.Column('id', sa.Uuid(), nullable=False),
+    sa.Column('sender_id', sa.Uuid(), nullable=False),
+    sa.Column('conversation_id', sa.Uuid(), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.Column('updated_at', sa.DateTime(), nullable=True),
+    sa.Column('deleted', sa.Boolean(), nullable=False),
+    sa.Column('deleted_at', sa.DateTime(), nullable=True),
+    sa.ForeignKeyConstraint(['conversation_id'], ['conversation.id'], ),
+    sa.ForeignKeyConstraint(['sender_id'], ['user.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('userblock',
+    sa.Column('blocker_id', sa.Uuid(), nullable=False),
+    sa.Column('blocked_id', sa.Uuid(), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.ForeignKeyConstraint(['blocked_id'], ['user.id'], ),
+    sa.ForeignKeyConstraint(['blocker_id'], ['user.id'], ),
+    sa.PrimaryKeyConstraint('blocker_id', 'blocked_id')
+    )
+    op.create_table('image',
+    sa.Column('filename', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=False),
+    sa.Column('file_path', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=False),
+    sa.Column('file_type', sa.Enum('JPEG', 'JPG', 'PNG', 'WEBP', 'GIF', name='imagefiletype'), nullable=False),
+    sa.Column('file_size', sa.Integer(), nullable=False),
+    sa.Column('is_primary', sa.Boolean(), nullable=False),
+    sa.Column('display_order', sa.Integer(), nullable=False),
+    sa.Column('id', sa.Uuid(), nullable=False),
+    sa.Column('listing_id', sa.Uuid(), nullable=False),
+    sa.ForeignKeyConstraint(['listing_id'], ['listing.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('leaseagreement',
+    sa.Column('filename', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=False),
+    sa.Column('file_path', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=False),
+    sa.Column('file_type', sa.Enum('PDF', 'TXT', name='leasefiletype'), nullable=False),
+    sa.Column('file_size', sa.Integer(), nullable=False),
+    sa.Column('description', sqlmodel.sql.sqltypes.AutoString(length=500), nullable=True),
+    sa.Column('id', sa.Uuid(), nullable=False),
+    sa.Column('listing_id', sa.Uuid(), nullable=False),
+    sa.ForeignKeyConstraint(['listing_id'], ['listing.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('listing_id')
+    )
+    op.create_table('messagereadstatus',
+    sa.Column('message_id', sa.Uuid(), nullable=False),
+    sa.Column('user_id', sa.Uuid(), nullable=False),
+    sa.Column('read_at', sa.DateTime(), nullable=False),
+    sa.ForeignKeyConstraint(['message_id'], ['message.id'], ),
+    sa.ForeignKeyConstraint(['user_id'], ['user.id'], ),
+    sa.PrimaryKeyConstraint('message_id', 'user_id')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('messagereadstatus')
+    op.drop_table('leaseagreement')
+    op.drop_table('image')
+    op.drop_table('userblock')
+    op.drop_table('message')
+    op.drop_table('listing')
+    op.drop_table('item')
+    op.drop_table('conversationparticipant')
+    op.drop_index(op.f('ix_user_phone_number'), table_name='user')
+    op.drop_index(op.f('ix_user_email'), table_name='user')
+    op.drop_table('user')
+    op.drop_table('conversation')
+    # ### end Alembic commands ###
