--- backend/app/alembic/versions/d4833295feb2_merging_listings.py
@@ -1,30 +1,0 @@
-"""merging listings
-
-Revision ID: d4833295feb2
-Revises: e1c8832e3dba
-Create Date: 2025-04-03 18:59:08.832549
-
-"""
-from typing import Sequence, Union
-import sqlmodel.sql.sqltypes
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'd4833295feb2'
-down_revision: Union[str, None] = 'e1c8832e3dba'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
