--- backend/app/alembic/versions/df28aa0800de_made_column_auto_logout_non_nullable.py
@@ -1,34 +1,0 @@
-"""Made column auto_logout non-nullable
-
-Revision ID: df28aa0800de
-Revises: 950410132acd
-Create Date: 2025-03-04 20:12:51.301101
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'df28aa0800de'
-down_revision: Union[str, None] = '950410132acd'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('user', 'auto_logout',
-               existing_type=sa.INTEGER(),
-               nullable=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('user', 'auto_logout',
-               existing_type=sa.INTEGER(),
-               nullable=True)
-    # ### end Alembic commands ###
