--- backend/app/alembic/versions/df4f0bb192ae_add_table_listing_s.py
@@ -1,46 +1,0 @@
-"""Add table listing(s)
-
-Revision ID: df4f0bb192ae
-Revises: 510c0f972c43
-Create Date: 2025-03-15 20:00:42.149541
-
-"""
-from typing import Sequence, Union
-import sqlmodel.sql.sqltypes
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = 'df4f0bb192ae'
-down_revision: Union[str, None] = '510c0f972c43'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('listing',
-    sa.Column('num_bedrooms', postgresql.JSONB(astext_type=sa.Text()), nullable=True),
-    sa.Column('num_bathrooms', postgresql.JSONB(astext_type=sa.Text()), nullable=True),
-    sa.Column('address', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=True),
-    sa.Column('realty_company', sqlmodel.sql.sqltypes.AutoString(length=255), nullable=True),
-    sa.Column('min_price', sa.Float(), nullable=True),
-    sa.Column('max_price', sa.Float(), nullable=True),
-    sa.Column('included_utilities', postgresql.JSONB(astext_type=sa.Text()), nullable=True),
-    sa.Column('security_deposit', sqlmodel.sql.sqltypes.AutoString(), nullable=True),
-    sa.Column('amenities', postgresql.JSONB(astext_type=sa.Text()), nullable=True),
-    sa.Column('lease_start_date', sqlmodel.sql.sqltypes.AutoString(), nullable=True),
-    sa.Column('lease_end_date', sqlmodel.sql.sqltypes.AutoString(), nullable=True),
-    sa.Column('id', sa.Uuid(), nullable=False),
-    sa.Column('owner_id', sa.Uuid(), nullable=False),
-    sa.ForeignKeyConstraint(['owner_id'], ['user.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('listing')
-    # ### end Alembic commands ###
