--- backend/app/alembic/versions/e91f52f0e7c6_add_column_is_2fa_enabled_to_user.py
@@ -1,33 +1,0 @@
-"""Add column is_2fa_enabled to user
-
-Revision ID: e91f52f0e7c6
-Revises: 4b3272ada45a
-Create Date: 2025-03-04 01:31:41.297999
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-import sqlmodel.sql.sqltypes
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'e91f52f0e7c6'
-down_revision: Union[str, None] = '4b3272ada45a'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('user', sa.Column('is_2fa_enabled', sa.Boolean(), nullable=True))
-    op.add_column('user', sa.Column('latest_otp', sqlmodel.sql.sqltypes.AutoString(), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('user', 'latest_otp')
-    op.drop_column('user', 'is_2fa_enabled')
-    # ### end Alembic commands ###
