--- backend/app/alembic/versions/fd69a55b31ac_add_column_is_2fa_enabled_to_public_.py
@@ -1,30 +1,0 @@
-"""Add column is_2fa_enabled to public User model
-
-Revision ID: fd69a55b31ac
-Revises: e91f52f0e7c6
-Create Date: 2025-03-04 02:24:50.412682
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'fd69a55b31ac'
-down_revision: Union[str, None] = 'e91f52f0e7c6'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
